removed styling in app.css. added styling to components in landing page. cards now show in a responsive grid.
