+++ social-cards/src/App.css
@@ -48,5 +48,9 @@
 body {
   margin: 0;
   padding: 0;
-  display: inline !important;
+}
+
+/* positions icons to the right side of the card */
+.follow-icons {
+  margin-left: 14vw;
 }
