+++ social-cards/src/components/LandingPage.jsx
@@ -1,6 +1,7 @@
 import { useState, useEffect } from 'react';
 import axios from 'axios';
 import PreviewCard from './PreviewCard';
+import FollowUser from './FollowUser';
 
 const LandingPage = ({ isAuthenticated, token }) => {
 
@@ -17,20 +18,29 @@ const LandingPage = ({ isAuthenticated, token }) => {
 
     return (
         <>
-            <div>
+            <div className='flex flex-wrap flex-col md:flex-row gap-4'>
                 {cards.map((card) => {
                     return (
-                        <PreviewCard
-                            key={card.id}
-                            font={card.font}
-                            frontText={card.front_text}
-                            backText={card.back_text}
-                            backgroundColor={card.background_color}
-                            uploadImage={card.uploadImage}
-                            token={token}
-                            creator={card.creator}
-                            creatorID={card.creator_id}
-                        />
+                        <>
+                            <div className='cardContainer'>
+                                <FollowUser
+                                    token={token}
+                                    creator={card.creator}
+                                    creatorID={card.creator_id}
+                                />
+                                <PreviewCard
+                                    key={card.id}
+                                    font={card.font}
+                                    frontText={card.front_text}
+                                    backText={card.back_text}
+                                    backgroundColor={card.background_color}
+                                    uploadImage={card.uploadImage}
+                                    token={token}
+                                    creator={card.creator}
+                                    creatorID={card.creator_id}
+                                />
+                            </div>
+                        </>
                     );
                 })}
             </div>
